REF: Switch to traits for async encode and decode
